refactor: postLike 패키지 생성하여 분리
